refactor: add files module
Path classes better organized into own module separate from ast classes
